move chapters helpers to lib folder
